Factor out to be overriden functions into BasicPipeOperations class.
